Refactor: Corrected erroneous namespaces
