+++ Backend.Backup/BackupServiceCollectionExtensions.cs
@@ -1,6 +1,6 @@
+using Backend.Backup.Managers;
 using Backend.Backup.Providers;
 using Backend.Backup.Services;
-using Backend.Core.Providers;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Backend.Backup;
