+++ Backend.Backup/Services/WebRTCBackupDataChannelHandler.cs
@@ -1,6 +1,5 @@
 using System.Collections.Concurrent;
 using System.Text;
-using Backend.Backup.Services;
 using Microsoft.Extensions.DependencyInjection;
 using SIPSorcery.Net;
