+++ Backend.WebApi/Controllers/BackupController.cs
@@ -1,4 +1,6 @@
-using Backend.Backup;
+using Backend.Backup.Managers;
+using Backend.Backup.Providers;
+using Backend.Backup.Services;
 using Backend.Core;
 using Backend.Core.Infrastructure;
 using Backend.Core.Services;
